+++ safe-road-backend/module-learning/src/main/java/ru/itmo/saferoad/learning/service/UserTopicProgressService.java
@@ -15,7 +15,7 @@ public interface UserTopicProgressService {
 	@NonNull
 	UserTopicProgress upsertStatus(@NonNull Long userId, @NonNull Integer topicId, @NonNull ProgressStatus status);
 
-	long countCompletedTopicsInSection(@NonNull Long userId, @NonNull Integer id);
+	int countCompletedTopicsInSection(@NonNull Long userId, @NonNull Integer id);
 
 	void unlockTopic(@NonNull Long userId, @NonNull Topic topic);
